odoo/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
hiredis==3.1.0
redis==5.2.1
celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.7.0
django-taggit==6.1.0
psycopg2-binary==2.9.10
tblib==3.0.0
